refactor(#6b): split cli.py into a cli/ package (1812 -> ~25 line entrypoint)
Pure structural move. cli.py becomes a package: _base.py holds the warnings
setup + console + the click group (a leaf, no submodule imports); the 28
commands + the kraken subgroup move into six cohesive command modules
(run/diagnostics/kraken/reporting/maintenance/redeem) that register onto
via import side effect. __init__.py is a ~25-line entrypoint that binds main +
console + re-exports AuramaurBot and imports the submodules.

Preserved exactly: every command name, the kraken subgroup,
(entrypoint), and the test-patch contract — run/dust_exit now resolve
 at CALL time (lazy import) so patch('auramaur.cli.
AuramaurBot') still intercepts. The 6 command modules added to the
no-unbound-names guard (the console/log free-name trap).

Verified: --help shows the identical 28 commands + 6 kraken subcommands; commands
execute end-to-end; cli + full suite green (896 passed); ruff clean.
