+++ pyproject.toml
@@ -57,7 +57,8 @@ line-length = 100
 # import cycles; E402 (import-not-at-top) there is by design, not a defect.
 "auramaur/bot.py" = ["E402"]
 "auramaur/strategy/engine.py" = ["E402"]
-"auramaur/cli.py" = ["E402"]
+"auramaur/cli/_base.py" = ["E402"]
+"auramaur/cli/__init__.py" = ["E402"]
 # Tests and research scripts: inline semicolons (E702), local setup imports
 # (E402), and throwaway locals (F841) are idiomatic here and not worth churn.
 "tests/**" = ["E402", "E702", "F841"]
